requirements.txt
aiofiles==0.8.0
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
Babel==2.9.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
et-xmlfile==1.1.0
frozenlist==1.3.0
fuzzywuzzy==0.18.0
idna==3.3
multidict==6.0.2
numpy==1.22.4
openpyxl==3.0.10
pandas==1.4.2
pyenchant==3.2.2
pymongo==4.1.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
requests==2.28.0
six==1.16.0
urllib3==1.26.9
yarl==1.7.2
